[/lifetime] Fix test flake - rm check for Ref usage after delete()
Previous assumption was that attempting to extract a value from a `Ref` after `delete()`ing it would result in a null pointer value, but the behaviour here seems to be undefined.

Would be nice to find a way to prevent Ref usage after deletion, but that seems either impossible in Zig or above my pay grade.

Also, wrap v10 tests in a describe block to disambiguate test cases in the node test output.